Makes a lot of change!!
Moves the mods on number layer to the bottom row
Though I am still thinking if may be keeping them in the middle row
with the numbers as top row might be a better idea,

Thing against keeping number row in top row is that, there is no actual
advantage of keeping mods in the home row in num layer other than
getting it same with the multimedia layer, again there are different
ways to look at it and all the modifiers are available in home row

Adds a workspace layer to move between the workspaces, part of me got
irritated when using top row mods and trying to move between the layers
I will try to move between the workspaces now with bottom row to check
how it will go, for now I will leave it at that

The tap dance is such a nice feature, it can be used to group together
keys based on their meaning or usage inorder to decrease the usage. I
think I can use the tap dance to Capitalise the words lets think of it
later

Started using combos for tab and enter which seems to be a very good
idea, also used the saved up thumb key for the purpose of ctrl

The only not nice part of the keyboard right now is the alt key present
on the left pinkie, which kinda feels wrong, I kept it with the
anticipation of usage with workspaces with yabai or hyprland, but this
feels like premature optimization at this point, and most of the
workspace shortcuts I want seems to be already possible in the workspace
layer or the navigation layer, though I will keep it there for now

One thing I would like to try with that alt key would be decrease one
shot duration from 5s to 1s may be? or try implementing callum one shot
mods copying the code from dallius blog

I wish I could create a gaming layer but at this point it feels like too
much for my mind to keep up with and ingeneral I am not playing any
games right now to warrant spending time on it, and it feels better to
buy an xbox controller right now

Would like to add unicode mainly for using emojis in neovim, I thinking
solving this in neovim software level might be better idea than going to
the firmware level, other place I would like emojis would be at twitter
but yeah I dont want to work on emojis for that purpose alone, but
creating an extension to detect when you are trying to add emojis is a
good idea